+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## Release 0.2.0 (2022-02-01T15:01:46)
+* added tests to package overview page
+* added spacing
+* fixed grouping of versions and added sort to versions header
+* prettier format
+* added dependency metric
+
 ## Release 0.1.0 (2022-01-17T13:46:55)
 * format files
 * using thoth enpoints to check if package exists